modularized code
MongoDB API fucntions are outside routes
